LU-9538 utils: Tool for syncing file LSOM xattr
Add a helper tool for syncing file LSOM xattr.
Firstly, register a new changelog user:
lctl --device lustre-MDT0000 changelog_register

After perform some file operations on Lustre file system, run
this tool to sync file LSOM xattr:
llsom_sync -u cl1 -m lustre-MDT0000 /mnt/lustre

Reviewed-on: https://review.whamcloud.com/30124
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ libcfs/include/libcfs/util/parser.h
@@ -107,7 +107,7 @@ char *Parser_strarg(char *inp, const char *prompt, const char *deft,
 int Parser_arg2int(const char *inp, long *result, int base);
 
 /* Convert human readable size string to and int; "1k" -> 1000 */
-int Parser_size(int *sizep, char *str);
+int Parser_size(unsigned long *sizep, char *str);
 
 /* Convert a string boolean to an int; "enable" -> 1 */
 int Parser_bool(int *b, char *str);
